Konflux Onboarding Prerequisites
Add a Containerfile which packages the plugin manager into a container
image. The plugin manager is intended to be used as a utility within
the main Jenkins container image. This Containerfile packages the
plugin manager .jar into a runnable image for testing purposes. The
image also makes it simple to extract the plugin manager .jar by
placing it in a well-known location
(`/deployments/jenkins-plugin-manager.jar`)

Signed-off-by: Adam Kaplan <adam.kaplan@redhat.com>